backend/requirements.txt
fastapi
uvicorn
pydantic

diffusers
transformers
accelerate
Pillow
